requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.24.4
pandas==2.2.2
scikit-learn==1.1.3
sksurv==0.17.2
tqdm==4.66.2
faiss-cpu==1.7.4
timm==0.9.5
matplotlib==3.8.0
pyyaml==6.0.1
setuptools==65.5.0
six==1.16.0
sympy==1.11.1
argparse==1.4.0
jsonschema==4.19.0
torchmetrics==1.0.1
Pillow==10.0.0
mamba-ssm
ctranspath 